Pull request in my Fork

+++ src/main/java/pickyeater/basics/food/Ingredient.java
@@ -5,11 +5,10 @@ import java.util.List;
 /**
  * @author Claudio Di Maio
  */
-
 public interface Ingredient {
     Quantity getQuantity();
     String getName();
     Nutrients getNutrients();
-    double getPrice();
+    float getPrice();
     List<String> getTags();
 }
