+++ src/main/java/pickyeater/basics/food/Nutrients.java
@@ -3,17 +3,16 @@ package pickyeater.basics.food;
 /**
  * @author Claudio Di Maio
  */
-
 public interface Nutrients {
-    double getProteins();
-    double getSimpleCarbs();
-    double getComplexCarbs();
-    double getFibers();
-    double getSaturatedFats();
-    double getUnSaturatedFats();
-    double getTransFats();
-    double getAlcohol();
-    double getCarbs();
-    double getFats();
-    double getCalories();
+    float getProteins();
+    float getSimpleCarbs();
+    float getComplexCarbs();
+    float getFibers();
+    float getSaturatedFats();
+    float getUnSaturatedFats();
+    float getTransFats();
+    float getAlcohol();
+    float getCarbs();
+    float getFats();
+    float getCalories();
 }
