+++ src/main/java/pickyeater/basics/user/DailyProgresses.java
@@ -7,14 +7,11 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface DailyProgresses {
-
     void addEatenMeal(Meal meal);
     void removeEatenMeal(Meal meal);
     Nutrients getEatenNutrients();
     List<Meal> getEatenMeals();
-    void addBurnedCalories(int calories);
+    void addBurnedCalories(float calories);
     int getBurnedCalories();
     LocalDate getDate();
-
-
 }
