+++ src/main/java/pickyeater/builders/IngredientBuilder.java
@@ -2,7 +2,6 @@
 // Source code recreated from a .class file by IntelliJ IDEA
 // (powered by FernFlower decompiler)
 //
-
 package pickyeater.builders;
 
 import pickyeater.basics.food.Ingredient;
@@ -10,15 +9,10 @@ import pickyeater.basics.food.Nutrients;
 import pickyeater.basics.food.Quantity;
 
 public interface IngredientBuilder {
-    void setPrice(double price);
-
+    void setPrice(float price);
     void setName(String name);
-
     void addTags(String... tags);
-
     void setNutrients(Nutrients nutrients);
-
     Ingredient build();
-
     void setQuantity(Quantity quantity);
 }
